Move u2f-api to src/lib as outisde of src doesn't work
