+++ src/lib/u2f-api.js
@@ -1,10 +1,10 @@
+// Copyright 2014-2015 Google Inc. All rights reserved.
+
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file or at
+// https://developers.google.com/open-source/licenses/bsd
+
 /* eslint-disable */
-//Copyright 2014-2015 Google Inc. All rights reserved.
-
-//Use of this source code is governed by a BSD-style
-//license that can be found in the LICENSE file or at
-//https://developers.google.com/open-source/licenses/bsd
-
 /**
  * @fileoverview The U2F api.
  */
