+++ src/pages/FidoRegisterPage.jsx
@@ -4,7 +4,7 @@ import { Centerer } from '../components/Centerer';
 import { FidoRegisterView } from '../components/FidoRegisterView';
 
 import { AuthPlz } from '../AuthPlz';
-import { u2f } from '../../lib/u2f-api';
+import { u2f } from '../lib/u2f-api';
 
 class FidoRegisterPage extends React.Component {
